Aastu.mobile.g3.mihret.data layer

+++ mobile/AASTU-mobile-group-3/ecommerse_collab/lib/features/product/data/repositories/product_repository_impl.dart
@@ -71,7 +71,7 @@ class ProductRepositoryImpl extends ProductRepository{
   }
 
   @override
-  Future<Either<Failure, Product>> updateProduct({required String id, required String name, required int price, required String description}) async {
+  Future<Either<Failure, Product>> updateProduct({required String id, required String name, required double price, required String description}) async {
    try{
     return Right(await productRemoteDataSource.updateProduct(id: id, name: name, price: price, description: description));
